Redefine when a generation starts and ends

Not all people agree on when a generation started and ended. This is especially true with the most recent few generations. The years I have chosen come from Howe & Strauss' book. This function lets you redefine the start and end years of a generation to fit your personal tastes.

redefine_generation(generation, start, end)

Arguments

generation
a string, exactly matching a generation's name

start
an integer, representing the year you want generation to start at

end
an integer, representing thet year you want generation to end at.

Details

The changes here modify the underlying data and will be reflected in any subsequent reference to the generational data. You can view what the current dataset is with show_generations. Any change you make can be reset using reset_generations.

Examples

redefine_generation("Millennial", 1980, 1997)

#> Gen X is now from 1964 to 1979

#> Millennial is now from 1980 to 1997

#> Gen Z is now from 1998 to 2030

show_generations()

#>              name start  end
#> 1           Gen Z  1998 2030
#> 2      Millennial  1980 1997
#> 3           Gen X  1964 1979
#> 4          Boomer  1946 1964
#> 5          Silent  1929 1946
#> 6            G.I.  1908 1929
#> 7            Lost  1886 1908
#> 8      Missionary  1865 1886
#> 9     Progressive  1844 1865
#> 10         Gilded  1822 1844
#> 11 Transcendental  1794 1822
#> 12     Compromise  1773 1794
#> 13     Republican  1746 1773
#> 14        Liberty  1727 1746
#> 15      Awakening  1704 1727
#> 16  Enlightenment  1675 1704
#> 17       Glorious  1649 1675
#> 18       Cavalier  1621 1649
#> 19        Puritan  1594 1621
#> 20  Parliamentary  1569 1594
#> 21    Elizabethan  1542 1569
#> 22       Reprisal  1517 1542
#> 23    Reformation  1497 1517
#> 24       Humanist  1459 1497
#> 25     Aurthurian  1435 1459