Rename a generation

Not all people refer to generations the same way. The names I have chosen come from Howe & Strauss' book, though I have taken some liberty with the most recent four generations. This family of functions lets you rename a generation to fit your personal tastes.

rename_generation(old_name, new_name)

use_zoomer()

use_gen_z()

use_gen_y()

use_millennial()

use_13th()

use_gen_x()

use_baby_boom()

use_boomer()

Arguments

old_name
a string, exactly matching a generation's name

new_name
a string, representing what you want that generation to be called

Details

The main function is rename_generation. All others are shortcuts for some of the more common changes I expect people to make.

The changes here modify the underlying data and will be reflected in any subsequent reference to the generational data. You can view what the current dataset is with show_generations. Any change you make can be reset using reset_generations.

Examples

rename_generation("Gen Z", "Zoomer")

#> Gen Z has been renamed Zoomer

rename_generation("Boomer", "Baby Boom")

#> Boomer has been renamed Baby Boom

show_generations()

#>              name start  end
#> 1          Zoomer  1998 2030
#> 2      Millennial  1980 1997
#> 3           Gen X  1964 1979
#> 4       Baby Boom  1946 1963
#> 5          Silent  1929 1945
#> 6            G.I.  1908 1928
#> 7            Lost  1886 1907
#> 8      Missionary  1865 1885
#> 9     Progressive  1844 1864
#> 10         Gilded  1822 1843
#> 11 Transcendental  1794 1821
#> 12     Compromise  1773 1793
#> 13     Republican  1746 1772
#> 14        Liberty  1727 1745
#> 15      Awakening  1704 1726
#> 16  Enlightenment  1675 1703
#> 17       Glorious  1649 1674
#> 18       Cavalier  1621 1648
#> 19        Puritan  1594 1620
#> 20  Parliamentary  1569 1593
#> 21    Elizabethan  1542 1568
#> 22       Reprisal  1517 1541
#> 23    Reformation  1497 1516
#> 24       Humanist  1459 1496
#> 25     Aurthurian  1435 1458

use_gen_z()

#> Zoomer has been renamed Gen Z

use_zoomer()

#> Gen Z has been renamed Zoomer

use_gen_y()

#> Millennial has been renamed Gen Y

use_millennial()

#> Gen Y has been renamed Millennial

use_13th()

#> Gen X has been renamed 13th

use_gen_x()

#> 13th has been renamed Gen X

use_boomer()

#> Baby Boom has been renamed Boomer

use_baby_boom()

#> Boomer has been renamed Baby Boom